+++ Teensy/include/Sequence/sequence.hpp
@@ -12,8 +12,8 @@ class Sequence {
         void pop ();
         void reset ();
 
-        void run (float timer, float dt, Robot *robot);
-        void monitor (float timer, float dt, action_kind *kind, Kinetic *goal);
+        void run (unsigned long timer, unsigned long dt, Robot *robot);
+        void monitor (unsigned long timer, unsigned long dt, action_kind *kind, Kinetic *goal);
         void goToStep (int istep);
 
         // save the monitoring ie goals at each dt, actions, etc
